+++ map1/lib/main.dart
@@ -1,8 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:map1/announcementItem.dart';
-import 'package:map1/constant.dart';
-// import 'package:map1/animator.dart';
-import 'carousel.dart';
+import 'package:map1/screen/wrapper.dart';
 
 void main() => runApp(MyApp());
 
@@ -11,94 +8,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: MyHomePage(title: 'Flutter Demo Home Page'),
+      home: Wrapper(),
     );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-  final String title;
-
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      extendBodyBehindAppBar: true,
-      bottomNavigationBar: BottomAppBar(
-        color: Colors.transparent,
-        child: Container(
-          height: 70,
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: <Widget>[
-              IconButton(
-                icon: Icon(Icons.home),
-                onPressed: () {},
-              ),
-              IconButton(
-                icon: Icon(Icons.add),
-                onPressed: () {},
-              ),
-              IconButton(
-                icon: Icon(Icons.person),
-                onPressed: () {},
-              ),
-              IconButton(
-                icon: Icon(Icons.chat),
-                onPressed: () {},
-              ),
-            ],
-          ),
-          decoration: BoxDecoration(
-            color: Colors.white,
-            borderRadius: BorderRadius.only(topLeft: Radius.circular(15.0), topRight: Radius.circular(15.0))
-          ),
-        ),
-      ),
-        body: Column(
-      children: <Widget>[
-        Align(
-          alignment: Alignment.centerLeft,
-          child: Padding(
-            padding: const EdgeInsets.only(left: 20.0, top: 68),
-            child: Text(
-              "Trending",
-              style: heading,
-            ),
-          ),
-        ),
-        Carousel(),
-        Align(
-          alignment: Alignment.centerLeft,
-          child: Padding(
-            padding: const EdgeInsets.only(left: 20),
-            child: Text(
-              "Announcements",
-              style: heading,
-            ),
-          ),
-        ),
-        Container(
-          height: 241,
-          color: Colors.white,
-          child: ListView.builder(
-            scrollDirection: Axis.vertical,
-            itemCount: 4,
-            itemBuilder: (context, index) {
-              return AnnouncementItem(title: index.toString());
-            }
-          ),
-        )
-      ],
-    ));
   }
 }
