+++ map1/lib/model/news.dart
