+++ map1/lib/screen/wrapper.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+import 'home/home.dart';
+
+class Wrapper extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Home();
+  }
+}
